Улучшенная версия бэкпропа

+++ BackProp/ParityDataset.cs
@@ -9,10 +9,10 @@ namespace BackProp
 
 	public class ParityDataset
 	{
-		public double[][] features;
-		public double[] labels;
+		public float[][] features;
+		public float[] labels;
 
-		public ParityDataset(double[][] features, double[] labels)
+		public ParityDataset(float[][] features, float[] labels)
 		{
 			if (features.Length != labels.Length)
 				throw new ArgumentException("Features and labels must have the same length");
@@ -23,7 +23,7 @@ namespace BackProp
 
 		public long Count => features.Length;
 
-		public (double[] features, double labels) GetItem(long index)
+		public (float[] features, float labels) GetItem(long index)
 		{
 			if (index < 0 || index >= Count)
 				throw new IndexOutOfRangeException($"Index {index} is out of range for dataset with {Count} items");
@@ -34,15 +34,15 @@ namespace BackProp
 		// Создание датасета для задачи четности (XOR)
 		public static ParityDataset CreateXORDataset()
 		{
-			double[][] features = new double[][]
+			float[][] features = new float[][]
 			{
-			new double[] {0, 0},
-			new double[] {0, 1},
-			new double[] {1, 0},
-			new double[] {1, 1}
+			new float[] {0, 0},
+			new float[] {0, 1},
+			new float[] {1, 0},
+			new float[] {1, 1}
 			};
 
-			double[] labels = new double[]
+			float[] labels = new float[]
 			{
 				0,
 				1,
@@ -57,12 +57,12 @@ namespace BackProp
 		public static ParityDataset CreateNParityDataset(int numBits)
 		{
 			long numSamples = (long)Math.Pow(2, numBits);
-			double[][] features = new double[numSamples][];
-			double[] labels = new double[numSamples];
+			float[][] features = new float[numSamples][];
+			float[] labels = new float[numSamples];
 
 			for (long i = 0; i < numSamples; i++)
 			{
-				features[i] = new double[numBits];
+				features[i] = new float[numBits];
 				int sum = 0;
 
 				// Заполняем features битовым представлением числа i
@@ -71,7 +71,7 @@ namespace BackProp
 					// Используем принцип периодов как в вашем примере
 					long period = (long)Math.Pow(2, j + 1);
 					long halfPeriod = period / 2;
-					double value = ((i % period) >= halfPeriod) ? 1.0 : 0.0;
+					float value = ((i % period) >= halfPeriod) ? 1.0f : 0.0f;
 
 					features[i][j] = value;
 					sum += (int)value;
